+++ src/components/tripDetail/TripPlanList.tsx
@@ -1,50 +1,22 @@
-import React from "react";
-import AddPlaceButton from "./AddPlaceButton";
 import * as S from "../../styles/tripDetail/TripPlanList.style";
-import { DayPlan } from "../../assets/data/tripPlanData";
-import TripPlanPlaceItem from "./TripPlanPlaceItem";
+import { DayPlan as DayPlanType } from "../../assets/data/tripPlanData";
+import DayPlan from "./DayPlan";
 
 interface Props {
-  plan?: DayPlan[];
+  plan?: DayPlanType[];
   days?: number;
 }
 function TripPlanList({ days, plan = [] }: Props) {
-  const dateToString = (date: Date) => {
-    const days = ["일", "월", "화", "수", "목", "금", "토"];
-    return `${date.getUTCMonth() + 1}.${date.getUTCDate()}/${
-      days[date.getUTCDay()]
-    }`;
-  };
-
   return (
     <S.Container>
       <S.DayFilterButton onClick={() => {}}>전체 일정</S.DayFilterButton>
-      <S.EditButton>편집</S.EditButton>
       <S.PlaceListContainer>
         {plan.map((dayPlan) => (
-          <div>
-            <S.DayInfo>
-              Day {dayPlan.day}
-              <span>{dateToString(dayPlan.date)}</span>
-            </S.DayInfo>
-            <S.PlaceList>
-              {dayPlan.route.length === 0 ? (
-                <AddPlaceButton />
-              ) : (
-                <>
-                  {dayPlan.route.map((place, index) => (
-                    <TripPlanPlaceItem
-                      place={place}
-                      index={index}
-                      addPlaceButton={
-                        index === dayPlan.route.length - 1 && true
-                      }
-                    />
-                  ))}
-                </>
-              )}
-            </S.PlaceList>
-          </div>
+          <DayPlan
+            day={dayPlan.day}
+            date={dayPlan.date}
+            route={dayPlan.route}
+          />
         ))}
       </S.PlaceListContainer>
     </S.Container>
